fix: move common functions to utils module
