+++ rust/src/tests/mod.rs
@@ -18,26 +18,6 @@ pub mod chmod;
 pub mod posix_fallocate;
 pub mod utimensat;
 
-/// Wrapper for `fchmodat(None, path, mode, FchmodatFlags::FollowSymlink)`.
-pub fn chmod<P: ?Sized + nix::NixPath>(path: &P, mode: nix::sys::stat::Mode) -> nix::Result<()> {
-    nix::sys::stat::fchmodat(
-        None,
-        path,
-        mode,
-        nix::sys::stat::FchmodatFlags::FollowSymlink,
-    )
-}
-
-/// Wrapper for `fchmodat(None, path, mode, FchmodatFlags::NoFollowSymlink)`.
-pub fn lchmod<P: ?Sized + nix::NixPath>(path: &P, mode: nix::sys::stat::Mode) -> nix::Result<()> {
-    nix::sys::stat::fchmodat(
-        None,
-        path,
-        mode,
-        nix::sys::stat::FchmodatFlags::NoFollowSymlink,
-    )
-}
-
 /// A handy extention to std::os::unix::fs::MetadataExt
 trait MetadataExt: StdMetadataExt {
     /// Return the file's last accessed time as a `TimeSpec`, including
