+++ rust/src/tests/posix_fallocate.rs
@@ -10,7 +10,8 @@ use nix::{
 use crate::{
     runner::context::{FileType, SerializedTestContext},
     test::{FileSystemFeature, TestContext},
-    tests::{assert_ctime_changed, assert_ctime_unchanged, chmod},
+    tests::{assert_ctime_changed, assert_ctime_unchanged},
+    utils::chmod,
 };
 
 crate::test_case! {
