+++ rust/src/utils.rs
@@ -1,0 +1,19 @@
+/// Wrapper for `fchmodat(None, path, mode, FchmodatFlags::FollowSymlink)`.
+pub fn chmod<P: ?Sized + nix::NixPath>(path: &P, mode: nix::sys::stat::Mode) -> nix::Result<()> {
+    nix::sys::stat::fchmodat(
+        None,
+        path,
+        mode,
+        nix::sys::stat::FchmodatFlags::FollowSymlink,
+    )
+}
+
+/// Wrapper for `fchmodat(None, path, mode, FchmodatFlags::NoFollowSymlink)`.
+pub fn lchmod<P: ?Sized + nix::NixPath>(path: &P, mode: nix::sys::stat::Mode) -> nix::Result<()> {
+    nix::sys::stat::fchmodat(
+        None,
+        path,
+        mode,
+        nix::sys::stat::FchmodatFlags::NoFollowSymlink,
+    )
+}
